+++ patches/src/main/kotlin/app/morphe/patches/shared/misc/gms/Fingerprints.kt
@@ -14,6 +14,13 @@ internal object GooglePlayUtilityFingerprint : Fingerprint(
     )
 )
 
+internal object IsGooglePlayServicesAvailableFingerprint : Fingerprint(
+    name = "isGooglePlayServicesAvailable",
+    accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.STATIC),
+    returnType = "I",
+    parameters = listOf("Landroid/content/Context;", "I")
+)
+
 internal object ServiceCheckFingerprint : Fingerprint(
     accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.STATIC),
     returnType = "V",
@@ -38,3 +45,25 @@ internal object OriginalPackageNameExtensionFingerprint : Fingerprint(
         methodDef.name == "getOriginalPackageName" && classDef.type == EXTENSION_CLASS_DESCRIPTOR
     }
 )
+
+internal fun getMainOnCreateFingerprint(activityClassType: String, targetBundleMethod: Boolean = true): Fingerprint {
+    require(activityClassType.endsWith(';')) {
+        "Class type must end with a semicolon: $activityClassType"
+    }
+
+    val fullClassType = activityClassType.startsWith('L')
+
+    return Fingerprint(
+        returnType = "V",
+        parameters = if (targetBundleMethod) {
+            listOf("Landroid/os/Bundle;")
+        } else {
+            listOf()
+        },
+        custom = { method, classDef ->
+            method.name == "onCreate" &&
+                    if (fullClassType) classDef.type == activityClassType
+                    else classDef.type.endsWith(activityClassType)
+        }
+    )
+}