+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.14.0-dev.2](https://github.com/hoo-dles/morphe-patches/compare/v1.14.0-dev.1...v1.14.0-dev.2) (2026-03-16)
+
+
+### Features
+
+* **HelloChinese:** Add `Enable Premium` patch ([1827fab](https://github.com/hoo-dles/morphe-patches/commit/1827fab273d6b243e21f1b2f58995daa7f0bda6c))
+
+# [1.14.0-dev.1](https://github.com/hoo-dles/morphe-patches/compare/v1.13.1-dev.1...v1.14.0-dev.1) (2026-03-16)
+
+
+### Features
+
+* **Windy:** Add `Enable Premium` patch ([76e4eb6](https://github.com/hoo-dles/morphe-patches/commit/76e4eb69312ece6f53d4d0c7c044697e96988703))
+
+## [1.13.1-dev.1](https://github.com/hoo-dles/morphe-patches/compare/v1.13.0...v1.13.1-dev.1) (2026-03-10)
+
+
+### Bug Fixes
+
+* **Teuida:** Update patches to latest (v1.21.16) ([efba5b2](https://github.com/hoo-dles/morphe-patches/commit/efba5b276d0713e3efc108f706d1f46cd9d67b5e))
+
 # [1.13.0](https://github.com/hoo-dles/morphe-patches/compare/v1.12.0...v1.13.0) (2026-03-10)